Update README.md
This project focuses on creating a Spring Boot application to manage persons and their parent-child relationships, leveraging Docker for containerization and Ngrok for making the local deployment accessible over the internet. The application uses a PostgreSQL database for persistence and follows a multi-layered architecture consisting of web, application, and persistence layers.